Adding travisci support for Pg 10
Note that this means other things have to be installed and the travis
user must be created.  The security considerations here are different
than in other cases.  However this is a minimal working examople that
actually runs.

Note that there are some tests that are not passing yet due to different
behavior in what is echoed.