Version 0.1.5, compatibility with PostgreSQL 11. (#13)
* Version 0.1.5, compatibility with PostgreSQL 11.

In PostgreSQL 11 there are built-in functions sha224, sha384, sha512
that accept bytea and return bytea. Since PostgreSQL only uses the
input parameter to resolve overloaded definitions, they conflict
with the cast functions from bytea provided by this extension. In
order to resolve this, case functions has been suffixed with '-b',
i.e. sha224b, sha512b etc (the same way as we already did with md5b.

Fix the parallel exection regression tests to match the output of
PostgreSQL 11.

Provide an upgrade script for 0.1.4 to 0.1.5 (basically renaming all
the functions involved). Also commit 0.1.4 SQL files, so that we
know what exactly is running on the hosts.

Since versions before 0.1.5 are not compatible with PostgreSQL 11
the upgrade test for 11+ is disabled.

* Add 0.1.5 generated file.

* Changes for 0.1.15 (leave only one changelog).

* Fix bytecode complication on llvm-enabled PG 11+.

For each files from the extension pgxs infrastructure attempts to
build a bytecode from the corresponding C source. It doesn't always
work, since the rules in Makefile.global discard all CFLAGS, therefore
in our case we need to passt the type name via the CPPFLAGS that are
honoted by the install_llvm_module.

* Remove an extra newline